Bugfix: Prevent sorting of unallocated metalink resource/metaurl
* src/main.c (main): Sort metalink resources only if there is more
  than 1 real resource allocated
* src/http.c (metalink_from_http): Sort metalink resources only if
  there is more than 1 real resource allocated
* src/http.c (metalink_from_http): Sort metalink metaurls only if
  there is more than 1 real metaurl allocated

Trying to sort the resources/metaurls of a metalink when there is only
one element allocated and no tail, or no elements at all, will lead to
a segmentation fault.

Also, sorting the resources/metaurls of a metalink when there is only
one element allocated and a NULL tail is useless.